V10
Il reste à faire la partie d'asservissement.

+++ Robot_Minh_Theo.X/robot.h
@@ -20,17 +20,17 @@ typedef struct robotStateBITS {
             float distanceTelemetrePlusDroit;
             unsigned int mode;
             
-            double vitesseDroitFromOdometry;
-            double vitesseGaucheFromOdometry;
-            double vitesseLineaireFromOdometry;
-            double vitesseAngulaireFromOdometry;
-            double xPosFromOdometry_1;
-            double xPosFromOdometry;
-            double yPosFromOdometry_1;
-            double yPosFromOdometry;
-            double vitesseBITEAngulaireFromOdometry;
-            double angleRadianFromOdometry_1;
-            double angleRadianFromOdometry;
+            float vitesseDroitFromOdometry;
+            float vitesseGaucheFromOdometry;
+            float vitesseLineaireFromOdometry;
+            float vitesseAngulaireFromOdometry;
+            float xPosFromOdometry_1;
+            float xPosFromOdometry;
+            float yPosFromOdometry_1;
+            float yPosFromOdometry;
+            float vitesseBITEAngulaireFromOdometry;
+            float angleRadianFromOdometry_1;
+            float angleRadianFromOdometry;
             
             PidCorrector PidX;
             PidCorrector PidTheta;
